refactor(emulator): Service層を撤去しhandler/storeに分離、ライブラリ採用で簡素化
ユーザー指摘を反映した大規模リファクタ:
- Service 構造体は「ドメインがないのに無理に挟むのは不自然」との指摘を受け撤去。
  Handler 構造体が Store と JWT 設定を直接保持する素直な構成にした
- Handler は HTTP プレゼンテーション層に専念し internal/auth/handler/ に分離
- Store は永続化層として internal/auth/store/ に分離（user / refresh / clone / password / model）
- 自前実装の internal/jwt は廃止し github.com/golang-jwt/jwt/v5 を直接使用
  （WithTimeFunc で注入 clock を渡すラッパだけ handler/jwt.go に残す）
- 自前実装の internal/uuidx は廃止し github.com/google/uuid を直接使用
- 固定 JWT (anon/service_role 用) は config/keys.go へ集約
- WHAT コメントを削除し WHY のみ残す方針で全体整理
- 旧 service_test.go (600+ 行) を store_test.go / handler_test.go に分割

Go 全テスト (race 検証込み) と application 結合テスト 72 件の合格を確認。
