+++ emulator/supabase/go.mod
@@ -3,6 +3,7 @@ module github.com/geek-teck-mentors/trend-diary/emulator/supabase
 go 1.24
 
 require (
-	github.com/golang-jwt/jwt/v5 v5.3.1 // indirect
-	golang.org/x/crypto v0.31.0 // indirect
+	github.com/golang-jwt/jwt/v5 v5.3.1
+	github.com/google/uuid v1.6.0
+	golang.org/x/crypto v0.31.0
 )
