+++ emulator/supabase/internal/auth/handler/emulator.go
@@ -1,19 +1,20 @@
-package auth
+package handler
 
 import (
 	"net/http"
 	"strings"
 
+	"github.com/geek-teck-mentors/trend-diary/emulator/supabase/internal/auth/store"
 	"github.com/geek-teck-mentors/trend-diary/emulator/supabase/internal/httpx"
 )
 
-func (s *Service) handleEmulatorReset(w http.ResponseWriter, _ *http.Request) {
-	s.store.Reset()
+func (h *Handler) handleEmulatorReset(w http.ResponseWriter, _ *http.Request) {
+	h.store.Reset()
 	w.WriteHeader(http.StatusNoContent)
 }
 
-func (s *Service) handleEmulatorSnapshot(w http.ResponseWriter, _ *http.Request) {
-	httpx.WriteJSON(w, http.StatusOK, s.store.Snapshot())
+func (h *Handler) handleEmulatorSnapshot(w http.ResponseWriter, _ *http.Request) {
+	httpx.WriteJSON(w, http.StatusOK, h.store.Snapshot())
 }
 
 type seedUserRequest struct {
@@ -21,8 +22,7 @@ type seedUserRequest struct {
 	Password string `json:"password"`
 }
 
-// handleEmulatorSeedUser はテストヘルパ用に password ハッシュ込みでユーザを直接登録する。
-func (s *Service) handleEmulatorSeedUser(w http.ResponseWriter, r *http.Request) {
+func (h *Handler) handleEmulatorSeedUser(w http.ResponseWriter, r *http.Request) {
 	var req seedUserRequest
 	if err := httpx.ReadJSON(r, &req); err != nil {
 		writeAPIError(w, http.StatusBadRequest, "invalid request body")
@@ -33,12 +33,12 @@ func (s *Service) handleEmulatorSeedUser(w http.ResponseWriter, r *http.Request)
 		writeAPIError(w, http.StatusBadRequest, "email and password are required")
 		return
 	}
-	hash, err := HashPassword(req.Password)
+	hash, err := store.HashPassword(req.Password)
 	if err != nil {
 		writeAPIError(w, http.StatusInternalServerError, err.Error())
 		return
 	}
-	u, err := s.store.CreateUser(req.Email, hash)
+	u, err := h.store.CreateUser(req.Email, hash)
 	if err != nil {
 		writeAPIError(w, http.StatusConflict, err.Error())
 		return
