+++ emulator/supabase/internal/auth/store/password.go
@@ -1,13 +1,12 @@
-package auth
+package store
 
 import "golang.org/x/crypto/bcrypt"
 
-// HashPassword は bcrypt でパスワードをハッシュ化する。コスト最小に近い値でテスト速度を確保。
+// MinCost を使うのはテスト/エミュレータ用途で本物の Supabase より弱いセキュリティで構わないため。
 func HashPassword(pw string) ([]byte, error) {
 	return bcrypt.GenerateFromPassword([]byte(pw), bcrypt.MinCost)
 }
 
-// VerifyPassword は bcrypt ハッシュとの一致を返す。
 func VerifyPassword(hash []byte, pw string) bool {
 	return bcrypt.CompareHashAndPassword(hash, []byte(pw)) == nil
 }
