+++ emulator/supabase/internal/auth/store/password_test.go
@@ -1,4 +1,4 @@
-package auth
+package store
 
 import "testing"
